chore: DCMAW-19786 PR comments
